Site Editor: Template list add rename action
